[Refactoring]Define mem module
