+++ common/src/lib.rs
@@ -5,6 +5,7 @@
 
 pub mod constant;
 pub mod debug;
+pub mod mem;
 pub mod size;
 pub mod vram;
 
@@ -14,34 +15,15 @@ extern crate x86_64;
 use crate::constant::INIT_RSP;
 use core::ptr;
 use size::{Byte, Size};
-use uefi::table::boot;
-
-#[repr(C)]
-pub struct MemMap {
-    ptr: *mut boot::MemoryDescriptor,
-    num_descriptors: usize,
-}
-
-impl MemMap {
-    pub fn new_from_slice(map: &mut [boot::MemoryDescriptor]) -> Self {
-        let ptr = map.as_mut_ptr();
-        let num_descriptors = map.len();
-
-        Self {
-            ptr,
-            num_descriptors,
-        }
-    }
-}
 
 #[repr(C)]
 pub struct BootInfo {
     vram_info: vram::Info,
-    mem_map: MemMap,
+    mem_map: mem::Map,
 }
 
 impl BootInfo {
-    pub fn new(vram_info: vram::Info, mem_map: MemMap) -> Self {
+    pub fn new(vram_info: vram::Info, mem_map: mem::Map) -> Self {
         Self { vram_info, mem_map }
     }
